# swagger.yaml
openapi: 3.0.0
info:
  title: HDR UK API 
  description: API for Tools and artefacts repository.
  version: 0.0.1
servers:
  - url: https://api.www.healthdatagateway.org/api
  - url: http://localhost:3001/api

  - url: https://api.{environment}.healthdatagateway.org:{port}/api
    variables:
      environment:
        default: latest
        description: The Environment name.
      port:
        enum:
          - '443'
        default: '443'
paths:
  /v1/datasets/{datasetID}:
    get:
      summary: Returns Dataset object.
      tags:
        - Datasets
      parameters:
        - in: path
          name: datasetID
          required: true
          description: The ID of the datset
          schema:
            type : string
            example: '756daeaa-6e47-4269-9df5-477c01cdd271'
      responses: 
        '200':
          description: OK    
  /v1/datasets/:
    get:
      summary: Returns List of Dataset objects.
      tags:
        - Datasets
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/datasets/?search={searchTerm}:
    get:
      summary: Returns dataset search results by searchTerm
      tags:
        - Datasets
      parameters:
        - name: searchTerm
          in: path
          required: true
          description: A search term.
          schema:
            type : string
            example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/projects/:
    post:
      summary: Returns a Project object with ID.
      security:
        - cookieAuth: []
      tags:
        - Projects
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                type:
                  type: string
                name:
                  type: string
                link:
                  type: string
                description:
                  type: string
                categories:
                  type: object
                  properties:
                    category:
                      type: string
                    programmingLanguage:
                      type: array
                      items:
                        type: string
                    programmingLanguageVersion:
                      type: string
                licence:
                  type: string
                authors:
                  type: array
                  items:
                    type: number
                tags:
                  type: object
                  properties:
                    features:
                      type: array
                      items:
                        type: string
                    topics:
                      type: array
                      items:
                        type: string      
              example:   # Sample object
                type: 'project'
                name: 'Epilepsy data research'
                link: 'http://epilepsy.org'
                description: 'Epilespy data research description'
                categories: { category: 'API', programmingLanguage: ['Javascript'], programmingLanguageVersion: '0.0.0'}
                licence: 'MIT licence'
                authors: [4495285946631793]
                tags: {features: ['Arbitrage'], topics: ['Epilepsy']}
      responses: 
        '200':
          description: OK
          
    get:
      summary: Returns List of Project objects.
      tags:
        - Projects
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
  /v1/projects/{id}:
    get:
      summary: Returns Project object.
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type : integer
            example: 441788967946948
      responses: 
        '200':
          description: OK
    patch:
      summary: Change status of the Project object.
      security:
        - cookieAuth: []
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 662346984100503
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                activeflag:
                  type: string
              example:   # Sample object
                activeflag: 'active'
      responses: 
        '200':
          description: OK
    put:
      summary: Returns edited Project object.
      security:
        - cookieAuth: []
      tags:
        - Projects
      parameters:
      - in: path
        name: id
        required: true
        description: The ID of the project
        schema:
          type : integer
          format: int64
          example: 26542005388306332
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                id:
                  type: number
                type:
                  type: string
                name:
                  type: string
                link:
                  type: string
                description:
                  type: string
                categories:
                  type: object
                  properties:
                    category:
                      type: string
                    programmingLanguage:
                      type: array
                      items:
                        type: string
                    programmingLanguageVersion:
                      type: string
                licence:
                  type: string
                authors:
                  type: array
                  items:
                    type: number
                tags:
                  type: object
                  properties:
                    features:
                      type: array
                      items:
                        type: string
                    topics:
                      type: array
                      items:
                        type: string    
                toolids:
                  type: array
                  items:
                    type: string
              example:   # Sample object
                id: 26542005388306332
                type: 'project'
                name: 'Research Data TEST EPILEPSY'
                link: 'http://localhost:8080/epilepsy'
                description: 'Epilespy data research description'
                categories: { category: 'API', programmingLanguage: ['Javascript'], programmingLanguageVersion: '1.0.0'}
                licence: 'MIT licence'
                authors: [4495285946631793]
                tags: {features: ['Arbitrage'], topics: ['Epilepsy']}
                toolids: []
      responses: 
        '200':
          description: OK
...